Split out DspComparable into several files.
